pyproject.toml
[project]
name = "link-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.40.0",
    "onnx>=1.21.0",
    "onnxruntime>=1.24.4",
    "onnxruntime-tools>=1.7.0",
    "optimum[onnxruntime]>=2.1.0",
    "psutil>=6.1.0",
    "pyyaml>=6.0.2",
    "sacrebleu>=2.5.1",
    "sentencepiece>=0.2.1",
    "transformers>=4.36,<4.58",
]
